refactor(gsd): extract auto-worktree creation module
